Use terraform values for redirect rules
This patch removes the custom file format for specifying redirect rules
and replaces it with Terraform values. The benefits are this that this
is more familiar to people using Terraform, it allows us to easily type
check the rules, and it removes the need for custom parsing code. The
downside is that redirect rules are now a bit more verbose and difficult
to read for humans.